Cast to enough size type before shifting

+++ lib/str.c
@@ -48,7 +48,7 @@ grn_str_charlen_utf8(grn_ctx *ctx, const unsigned char *str, const unsigned char
   if (*str & 0x80) {
     int i;
     int len;
-    GRN_BIT_SCAN_REV(~(*str << 24), len);
+    GRN_BIT_SCAN_REV(~(((uint32_t)*str) << 24), len);
     len = 31 - len;
     if ((unsigned int)(len - 2) >= 3) {  /* (len == 1 || len >= 5) */
       GRN_LOG(ctx, GRN_LOG_WARNING,
